implement authentication features with JWT, including user registration and login; add validation for player and team operations

+++ src/main/java/com/palombetas/api/gerartimes/domain/dto/response/PlayerResponseDTO.java
@@ -5,7 +5,7 @@ import com.palombetas.api.gerartimes.domain.entity.PlayerEntity;
 public record PlayerResponseDTO(
         Long id,
         String name,
-        Double rating,
+        Integer rating,
         Integer timesChampion,
         Integer timesMvp
 ) {
